Rename scripts to kebab case
