+++ package.json
@@ -69,10 +69,10 @@
     "build": "yarn copy-contributors && yarn build:app && yarn build:lambda",
     "build:app": "gatsby build",
     "build:lambda": "netlify-lambda build src/lambda",
-    "copy-contributors": "node src/scripts/copyContributors.js",
+    "copy-contributors": "node src/scripts/copy-contributors.js",
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
-    "generate-heading-ids": "node src/scripts/generateHeadingIDs.js",
-    "merge-translations": "node src/scripts/mergeTranslations.js",
+    "generate-heading-ids": "node src/scripts/generate-heading-ids.js",
+    "merge-translations": "node src/scripts/merge-translations.js",
     "start": "gatsby develop",
     "start:static": "gatsby build && gatsby serve",
     "serve": "gatsby serve",
