+++ src/scripts/copy-contributors.js
