+++ src/scripts/merge-translations.js
